refactor: created user directory and add users
